Moving counter support to filebeat
The way the counter support is setup is very specific for filebeat
and how Filebeat hooks up a global ACKer, and the Registrar for storing
file.State. Moving it to filebeat to not encourage someone to use a it.
